Chore: Fixed some compiler warnings

+++ src/test/value.cxx
@@ -16,11 +16,11 @@ TEST(ValueTest, InvalidConstructor) {
 
 TEST(ValueTest, Equality) {
   std::vector<atn::sudoku::Value> values = { {}, {4}, {2}, {4} };
-  for (uint8_t i{0}; i < values.size(); ++i) {
+  for (u_int i{0}; i < values.size(); ++i) {
     EXPECT_TRUE(values[i] == values[i]);
   }
-  for (uint8_t i{0}; i < values.size() - 1; ++i) {
-    for (uint8_t j{i + 1}; j < values.size(); ++j) {
+  for (u_int i{0}; i < values.size() - 1; ++i) {
+    for (u_int j{i + 1}; j < values.size(); ++j) {
       if (i == 1 && j == 3) {
         EXPECT_TRUE(values[i] == values[j]);
       } else {
@@ -32,11 +32,11 @@ TEST(ValueTest, Equality) {
 
 TEST(ValueTest, Inequality) {
   std::vector<atn::sudoku::Value> values = { {}, {4}, {2}, {4} };
-  for (uint8_t i{0}; i < values.size(); ++i) {
+  for (u_int i{0}; i < values.size(); ++i) {
     EXPECT_FALSE(values[i] != values[i]);
   }
-  for (uint8_t i{0}; i < values.size() - 1; ++i) {
-    for (uint8_t j{i + 1}; j < values.size(); ++j) {
+  for (u_int i{0}; i < values.size() - 1; ++i) {
+    for (u_int j{i + 1}; j < values.size(); ++j) {
       if (i == 1 && j == 3) {
         EXPECT_FALSE(values[i] != values[j]);
       } else {
